333. Deleting Files: postDeleteProduct unlink prod images

+++ controllers/admin.js
@@ -1,4 +1,7 @@
 const mongoose = require('mongoose')
+
+const fileHelper = require('../utils/file');
+
 const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -173,6 +176,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       // As mongoose will return the monodb object itself with all its methods
@@ -216,9 +220,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      console.log('Deleted?');
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
-      console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
     .catch(err => {
